+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useLogin } from './useLogin';
 export { default as useToast } from './useToast';
+export { default as useModal } from './useModal';
